feat: data:timetable:view 세팅
