+++ data/timetable/viewer/build.gradle.kts
@@ -1,0 +1,8 @@
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
+plugins {
+    id("suwiki.android.data")
+}
+
+android {
+    namespace = "com.suwiki.data.timetable.viewer"
+}
