+++ data/timetable/viewer/proguard-rules.pro
