+++ data/timetable/viewer/src/main/AndroidManifest.xml
