+++ data/timetable/viewer/src/main/java/com/suwiki/data/timetable/viewer/datasource/LocalTimetableProviderDatasource.kt
@@ -1,0 +1,7 @@
+package com.suwiki.data.timetable.viewer.datasource
+
+import com.suwiki.core.model.timetable.TimetableData
+
+interface LocalTimetableProviderDatasource {
+    suspend fun getLocalTimetable(): List<TimetableData>
+}
