+++ data/timetable/viewer/src/main/java/com/suwiki/data/timetable/viewer/datasource/RemoteTimetableDataSource.kt
@@ -1,0 +1,8 @@
+package com.suwiki.data.timetable.viewer.datasource
+
+import com.suwiki.core.model.timetable.TimetableData
+
+interface RemoteTimetableDataSource {
+    suspend fun fetchRemoteTimetableVersion(): Long
+    suspend fun fetchRemoteTimetable(): List<TimetableData>
+}
