+++ local/timetable/viewer/src/main/java/com/suwiki/local/timetable/viewer/datasource/LocalTimetableProviderDatasourceImpl.kt
@@ -1,8 +1,8 @@
 package com.suwiki.local.timetable.viewer.datasource
 
-import com.suwiki.data.datasource.local.LocalTimetableProviderDatasource
 import com.suwiki.core.database.TimetableDatabase
 import com.suwiki.core.model.timetable.TimetableData
+import com.suwiki.data.timetable.viewer.datasource.LocalTimetableProviderDatasource
 import com.suwiki.local.timetable.viewer.converter.toModel
 import javax.inject.Inject
