+++ settings.gradle.kts
@@ -29,7 +29,7 @@ include(":di")
 include(":core:network")
 include(":core:security")
 include(":remote:openmajor")
-include(":remote:timetable:editor")
+include(":remote:timetable:viewer")
 include(":remote:lectureevaluation:viewer")
 include(":remote:lectureevaluation:report")
 include(":remote:lectureevaluation:my")
@@ -44,3 +44,4 @@ include(":local:timetable:editor")
 include(":local:user")
 include(":local:token:storage")
 include(":data:openmajor")
+include(":data:timetable:viewer")
